+++ src/main/java/btrplace/plan/action/ForgeVM.java
@@ -76,7 +76,9 @@ public class ForgeVM extends Action {
             return true;
         } else if (o.getClass() == this.getClass()) {
             ForgeVM that = (ForgeVM) o;
-            return this.id.equals(that.id);
+            return this.id.equals(that.id)
+                    && this.getStart() == that.getStart()
+                    && this.getEnd() == that.getEnd();
         }
         return false;
     }